requirements.txt
asgiref==3.7.2
astroid==2.15.6
certifi==2023.7.22
contourpy==1.1.1
coverage==7.3.2
cycler==0.11.0
dill==0.3.7
Django==4.2.6
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
filelock==3.13.1
fonttools==4.42.1
fsspec==2023.12.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
iniconfig==2.0.0
isort==5.12.0
jax==0.4.16
jaxlib==0.4.16
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mccabe==0.7.0
ml-dtypes==0.3.1
mpmath==1.3.0
mushroom-rl==1.10.0.post2
networkx==3.2.1
numpy==1.26.0
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.2
Pillow==10.0.1
platformdirs==3.10.0
pluggy==1.3.0
pygame==2.5.2
PyJWT==2.8.0
pylint==2.17.5
pyparsing==3.1.1
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.11.2
six==1.16.0
sqlparse==0.4.4
sympy==1.12
threadpoolctl==3.2.0
tomli==2.0.1
tomlkit==0.12.1
torch==2.1.1
tqdm==4.66.1
typing-extensions==4.8.0
tzdata==2023.3
wrapt==1.15.0
zipp==3.16.2
